services/auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.34
psycopg[binary]==3.2.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

pydantic==2.9.2
email-validator==2.2.0

pika==1.3.2

mangum==0.17.0